pyproject.toml
[tool.poetry]
name = "djangorestproject"
version = "0.1.0"
description = ""
authors = ["OlegMatus <129312840+OlegMatus@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.6"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
